Move the sample project to sub directory
