+++ Sample/SampleTargetProjects/dotnetConsole/.gitignore
