+++ Sample/SampleTargetProjects/dotnetConsole/SampleTargetProjects.csproj
