+++ Sample/SampleTargetProjects/localeFilenameTest/English.txt
