+++ Sample/SampleTargetProjects/localeFilenameTest/Français.txt
