+++ Sample/SampleTargetProjects/reactapp/LICENSE.txt
