+++ Sample/SampleTargetProjects/reactapp/public/favicon.ico
