+++ Sample/SampleTargetProjects/reactapp/public/manifest.json
