+++ Sample/SampleTargetProjects/reactapp/src/App.tsx
