+++ Sample/SampleTargetProjects/reactapp/src/index.tsx
