+++ Sample/SampleTargetProjects/reactapp/src/react-app-env.d.ts
