+++ Sample/SampleTargetProjects/reactapp/src/serviceWorker.ts
