+++ Sample/UnitTest/BasicFilesChecker.cs
