+++ Sample/UnitTest/CSharp/DefaultDesktopOnlyChecker.cs
