+++ Sample/UnitTest/TextEncodeChecker.cs
