+++ Sample/UnitTest/Web/DoctypeChecker.cs
