+++ Sample/UnitTest/Web/VarChecker.cs
